ensure list attribute changes trigger a cell model update
while the validation fired correctly without this change the Cell.update method wasn't being called at the right time. This led to cells not being removed from the set of invalid notes that the form uses to block submissions. This was only triggered if the member field was nonempty and the group field was changed from a group A to a group B where the member belonged to both groups